+++ README.md
@@ -9,4 +9,4 @@ Split the `App.js` into components based on CSS blocks. CSS code should be split
 - Install Prettier Extention and use this [VSCode settings](https://mate-academy.github.io/fe-program/tools/vscode/settings.json) to enable format on save.
 - Implement a solution following the [React task guideline](https://github.com/mate-academy/react_task-guideline#react-tasks-guideline)
 - Open one more terminal and run tests with `npm test` to ensure your solutions is correct
-- Replace `<your_account>` with your Github username in the [DEMO LINK](https://<your_account>.github.io/react_decompose/) and add it to the PR description
+- Replace `<your_account>` with your Github username in the [DEMO LINK](https://Varvara00.github.io/react_decompose/) and add it to the PR description
